reorder timestamps when merging
I was thinking that GTKWave could handle VCD files with out-of-order
timestamps, but this is not true, at least when appending to a really
big file.

This change now merge sort the two files. Futhermore, the code does not
assume that each file is already sorted, instead it find each sorted
section of timestamps for each file, and do a n-way merge of all of
them, using a binary heap.